Added ability to comment on requests.
This is the hackiest and dirtiest fix I've ever done.
I set the form to default to a rating of 1, changed the model of
requests to allow ratings, hid all of the rating forms, and just hide
the ratings from the display.

But from the user's perspective, it pretty much works, people *could*
put phoney ratings and stuff but it doesn't really matter.

Change-Id: I6fbaf30ab9539d40aa68b791167024c18af08fab